+++ bluetooth_low_energy_android/example/lib/views/central_manager_view.dart
@@ -132,7 +132,7 @@ class _CentralManagerViewState extends State<CentralManagerView> {
               trailing: Row(
                 mainAxisSize: MainAxisSize.min,
                 children: [
-                  RssiWidget(rssi),
+                  RssiIndicator(rssi),
                   Text('$rssi'),
                 ],
               ),