+++ bluetooth_low_energy_platform_interface/lib/src/central.dart
@@ -1,7 +1,7 @@
 import 'bluetooth_low_energy_peer.dart';
 
 /// A remote device connected to a local app, which is acting as a peripheral.
-abstract base class Central extends BluetoothLowEnergyPeer {
+base class Central extends BluetoothLowEnergyPeer {
   Central({
     required super.uuid,
   });