+++ bluetooth_low_energy_windows/lib/src/my_peripheral.dart
@@ -1,0 +1,13 @@
+import 'package:bluetooth_low_energy_platform_interface/bluetooth_low_energy_platform_interface.dart';
+
+import 'my_api.dart';
+
+base class MyPeripheral extends Peripheral {
+  final int address;
+
+  MyPeripheral({
+    required this.address,
+  }) : super(
+          uuid: address.toUUID(),
+        );
+}