+++ bluetooth_low_energy_windows/windows/bluetooth_low_energy_windows.cpp
@@ -8,7 +8,7 @@ namespace bluetooth_low_energy_windows {
 	void BluetoothLowEnergyWindows::RegisterWithRegistrar(flutter::PluginRegistrarWindows* registrar) {
 		auto messenger = registrar->messenger();
 		auto central_manager = std::make_unique<MyCentralManager>(messenger);
-		MyCentralManagerHostApi::SetUp(messenger, central_manager.get());
+		MyCentralManagerHostAPI::SetUp(messenger, central_manager.get());
 
 		auto plugin = std::make_unique<BluetoothLowEnergyWindows>(std::move(central_manager));
 		registrar->AddPlugin(std::move(plugin));