+++ bluetooth_low_energy_windows/windows/my_central_manager.h
@@ -10,7 +10,7 @@
 
 namespace bluetooth_low_energy_windows
 {
-	class MyCentralManager : public MyCentralManagerHostApi
+	class MyCentralManager : public MyCentralManagerHostAPI
 	{
 	public:
 		MyCentralManager(flutter::BinaryMessenger* messenger);
@@ -21,8 +21,8 @@ namespace bluetooth_low_energy_windows
 		MyCentralManager& operator=(const MyCentralManager&) = delete;
 
 		// 通过 MyCentralManagerHostApi 继承
-		void SetUp(std::function<void(std::optional<FlutterError>reply)> result) override;
-		std::optional<FlutterError> StartDiscovery() override;
+		std::optional<FlutterError> Initialize() override;
+		std::optional<FlutterError> StartDiscovery(const flutter::EncodableList& service_uuids_args) override;
 		std::optional<FlutterError> StopDiscovery() override;
 		void Connect(int64_t address_args, std::function<void(std::optional<FlutterError>reply)> result) override;
 		std::optional<FlutterError> Disconnect(int64_t address_args) override;
@@ -35,7 +35,7 @@ namespace bluetooth_low_energy_windows
 		void ReadDescriptor(int64_t address_args, int64_t handle_args, std::function<void(ErrorOr<std::vector<uint8_t>>reply)> result) override;
 		void WriteDescriptor(int64_t address_args, int64_t handle_args, const std::vector<uint8_t>& value_args, std::function<void(std::optional<FlutterError>reply)> result) override;
 	private:
-		std::optional<MyCentralManagerFlutterApi> m_api;
+		std::optional<MyCentralManagerFlutterAPI> m_api;
 		std::optional<winrt::Windows::Devices::Bluetooth::Advertisement::BluetoothLEAdvertisementWatcher> m_watcher;
 		std::optional<winrt::Windows::Devices::Bluetooth::BluetoothAdapter> m_adapter;
 		std::optional<winrt::Windows::Devices::Radios::Radio> m_radio;
@@ -48,7 +48,7 @@ namespace bluetooth_low_energy_windows
 		std::map<int64_t, std::optional<winrt::Windows::Devices::Bluetooth::BluetoothLEDevice::ConnectionStatusChanged_revoker>> m_device_connection_status_changed_revokers;
 		std::map<int64_t, std::map<int64_t, std::optional<winrt::Windows::Devices::Bluetooth::GenericAttributeProfile::GattCharacteristic::ValueChanged_revoker>>> m_characteristic_value_changed_revokers;
 
-		winrt::fire_and_forget m_set_up(std::function<void(std::optional<FlutterError>reply)> result);
+		winrt::fire_and_forget m_initialize();
 		winrt::fire_and_forget m_connect(int64_t address_args, std::function<void(std::optional<FlutterError>reply)> result);
 		winrt::fire_and_forget m_discover_services(int64_t address_args, std::function<void(ErrorOr<flutter::EncodableList> reply)> result);
 		winrt::fire_and_forget m_discover_characteristics(int64_t address_args, int64_t handle_args, std::function<void(ErrorOr<flutter::EncodableList> reply)> result);
@@ -59,8 +59,7 @@ namespace bluetooth_low_energy_windows
 		winrt::fire_and_forget m_read_descriptor(int64_t address_args, int64_t handle_args, std::function<void(ErrorOr<std::vector<uint8_t>> reply)> result);
 		winrt::fire_and_forget m_write_descriptor(int64_t address_args, int64_t handle_args, const std::vector<uint8_t>& value_args, std::function<void(std::optional<FlutterError> reply)> result);
 
-		void m_clear_state();
-		void m_clear_device(int64_t address_args);
+		void m_disconnect(int64_t address_args);
 		void m_on_state_changed();
 		MyBluetoothLowEnergyStateArgs m_radio_state_to_args(winrt::Windows::Devices::Radios::RadioState state);
 		MyAdvertisementArgs m_advertisement_to_args(const winrt::Windows::Devices::Bluetooth::Advertisement::BluetoothLEAdvertisement& advertisement);